[CMAT-50] feat:왜래키 관련 오류 수정

+++ src/main/java/UMC/career_mate/domain/planner/service/PlannerCommandService.java
@@ -39,6 +39,10 @@ public class PlannerCommandService {
 
     public void editPlanner(Member member, CreatePlannerListDTO createPlannerListDTO){
 
+        if(!plannerRepository.existsByMember(member)){
+            throw new GeneralException(CommonErrorCode.PLANNER_NOT_EXISTS);
+        }
+
         List<Planner> planners = plannerRepository.findByMember(member);
         List<CreatePlannerDTO> plannerDTOs = createPlannerListDTO.planners();
 
@@ -48,6 +52,10 @@ public class PlannerCommandService {
     }
 
     public void deletePlanner(Member member){
+
+        if(!plannerRepository.existsByMember(member)){
+            throw new GeneralException(CommonErrorCode.PLANNER_NOT_EXISTS);
+        }
 
         List<Planner> planners = plannerRepository.findByMember(member);
         plannerRepository.deleteAll(planners);
